+++ src/main/java/com/buuz135/replication/block/tile/ChipStorageBlockEntity.java
@@ -2,6 +2,7 @@ package com.buuz135.replication.block.tile;
 
 import com.buuz135.replication.ReplicationRegistry;
 import com.buuz135.replication.api.pattern.IMatterPatternHolder;
+import com.buuz135.replication.api.pattern.IMatterPatternModifier;
 import com.buuz135.replication.api.pattern.MatterPattern;
 import com.buuz135.replication.client.gui.ReplicationAddonProvider;
 import com.buuz135.replication.client.gui.addons.ChipStorageAddon;
@@ -14,6 +15,7 @@ import net.minecraft.core.Direction;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.ItemInteractionResult;
 import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockState;
@@ -82,6 +84,26 @@ public class ChipStorageBlockEntity extends NetworkBlockEntity<ChipStorageBlockE
 
     private void notifyNetworkOfSlotChange() {
         if (isServer()){
+            for (int i = 0; i < this.chips.getSlots(); i++) {
+                var stack = this.chips.getStackInSlot(i);
+                if (!stack.isEmpty() && stack.getItem() instanceof IMatterPatternHolder stackHolder) {
+                    List<MatterPattern> patterns = stackHolder.getPatterns(this.level, stack);
+                    for (int j = 0; j < this.chips.getSlots(); j++) {
+                        if (i == j) continue;
+                        var otherStack = this.chips.getStackInSlot(j);
+                        if (!stack.isEmpty() && otherStack.getItem() instanceof IMatterPatternHolder otherStackHolder && otherStack.getItem() instanceof IMatterPatternModifier otherPatternModifier) {
+                            List<MatterPattern> otherPatterns = otherStackHolder.getPatterns(this.level, otherStack);
+                            for (MatterPattern pattern : patterns) {
+                                for (MatterPattern otherPattern : otherPatterns) {
+                                    if (ItemStack.isSameItemSameComponents(pattern.getStack(), otherPattern.getStack())) {
+                                        otherPatternModifier.removePattern(this.level, otherStack, otherPattern.getStack());
+                                    }
+                                }
+                            }
+                        }
+                    }
+                }
+            }
             syncObject(this.chips);
             cachePatterns();
             this.getNetwork().onChipValuesChanged(this, this.worldPosition);